chore: rename interrface.tsx to interface.tsx
