+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { FC } from 'react';
-import { MainPageProps } from './interface/interrface';
+import { MainPageProps } from './interface/interface';
 import RoutRoot from './router/routerRoot';
 
 const App: FC<MainPageProps> = ({ offers }) => <RoutRoot offers={offers} />;
