+++ src/components/protectRoute.tsx
@@ -1,7 +1,7 @@
 import { Navigate, Outlet } from 'react-router-dom';
 import { AppRoute, AuthorizationStatus } from '../consts';
 import { FC } from 'react';
-import { AuthProps } from '../interface/interrface';
+import { AuthProps } from '../interface/interface';
 
 const ProtectRoute: FC<AuthProps> = ({ authorizationStatus }) =>
   authorizationStatus !== AuthorizationStatus.Auth ? (
